datatool: port away from assimp.
Using SceneTools to perform a transformation and concatenation of all
meshes in the file instead. The copy back to the ancient habitat
MeshData structure is what makes this longer than it could be but I
don't want to touch that as well now.

With this, the Assimp::Assimp CMake target doesn't need to be found (or
made) anymore, it was additionally errorneously linked to the bindings as
well.