Update Magnum to latest (#2521)
* datatool: make lowercase again.

Seems like an accident made in 94255f73b7db98844e66f86199b755129676107b
(September 2021).

* datatool: port away from assimp.

Using SceneTools to perform a transformation and concatenation of all
meshes in the file instead. The copy back to the ancient habitat
MeshData structure is what makes this longer than it could be but I
don't want to touch that as well now.

With this, the Assimp::Assimp CMake target doesn't need to be found (or
made) anymore, it was additionally errorneously linked to the bindings as
well.

* Update to latest Magnum.

The applications got an overhaul for first-class touch and pen support,
meaning the events are renamed now. Since GLFW doesn't support touch or
pen, nothing else needs to be done really.

Besides that, Magnum itself bundles the Find modules, meaning they don't
need to have copies here anymore, unless used directly such as
FindBullet.

* CMake: include 3rd party dependencies as SYSTEM to reduce warning spam.

* refactor viewer.py applications for new Application Pointer API updates

---------

Co-authored-by: Alexander Clegg <alexanderwclegg@gmail.com>