assets: contrary to the comment, this was mainly a workaround.
In some cases such as depth-only rendering it may be desirable to have
for example positions *not* interleaved with the remaining attributes,
and this would undo that optimization.

The code below however calls mutableIndices(), and without the
interleave it blows up because not all meshes have mutable index data,
apparently. So we need to expand that check there.