clang format wants to say something?!
The machine is so amazing, a human could never format it like this, it
would always have some silly slightly different opinion. Result? It
makes me *hate* the code I write because it's never good enough for the
machine, and I don't enjoy reading the formatted thing either.

We've achieved peak technology.