Update to latest Magnum with more efficient glTF mesh import and various UX improvements (#2551)
* Update to latest Magnum.

* assets: contrary to the comment, this was mainly a workaround.

In some cases such as depth-only rendering it may be desirable to have
for example positions *not* interleaved with the remaining attributes,
and this would undo that optimization.

The code below however calls mutableIndices(), and without the
interleave it blows up because not all meshes have mutable index data,
apparently. So we need to expand that check there.

* assets: can use this nice helper also.

Could also resolve the TODO there because generateIndices() is a thing
that works, but don't feel like doing that now.

* Use arrayClear() instead of arrayResize(..., 0).

* Use the path() and filename() shorthands for Path::split().

* clang format wants to say something?!

The machine is so amazing, a human could never format it like this, it
would always have some silly slightly different opinion. Result? It
makes me *hate* the code I write because it's never good enough for the
machine, and I don't enjoy reading the formatted thing either.

We've achieved peak technology.